:recycle: Rename generate-acknowledgement service folder to generate-confidentiality-acknowledgement.
